+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 ## 2026-05-27 - Git commit version stamp
 
+- Version: `0.1.9`
+- Timestamp: 2026-05-27 06:52 (UTC+7)
+- Commit: `de15859`
+- Type: Patch
+- Status: Draft
+
+### Changes
+
+- Version stamp for git commit.
+
+### Verification
+
+- pending
+
+---
+## 2026-05-27 - Git commit version stamp
+
 - Version: `0.1.8`
 - Timestamp: 2026-05-27 05:45 (UTC+7)
 - Commit: `a2907da`